Fix CI to not delete constants if it already exists

+++ ci.sh
@@ -2,8 +2,14 @@
 
 set -euo pipefail
 
-rm laundry-monitor/constants.h || true
-cp laundry-monitor/constants.sample.h laundry-monitor/constants.h
-platformio run
-rm laundry-monitor/constants.h
+constants_created=
+if [ ! -f "laundry-monitor/constants.h" ]; then
+  cp laundry-monitor/constants.sample.h laundry-monitor/constants.h
+  constants_created=true
+fi
 
+platformio run
+
+if [ "$constants_created" = true ]; then
+  rm laundry-monitor/constants.h
+fi
